Smart Binary Threshold Implemented + Cleaned Stuff
Looks promising, image 2 is actually kinda okay if kmeans=8 just press run and look
